Add script for testing
