+++ requirements.txt
@@ -1,14 +1,18 @@
 certifi==2019.11.28
 chardet==3.0.4
+cycler==0.10.0
 dominate==2.5.1
 idna==2.9
 joblib==0.14.1
 jsonpatch==1.25
 jsonpointer==2.0
+kiwisolver==1.1.0
+matplotlib==3.0.3
 numpy==1.18.1
 opencv-python==4.2.0.32
 pandas==0.25.3
 Pillow==7.0.0
+pyparsing==2.4.7
 python-dateutil==2.8.1
 pytz==2020.1
 pyzmq==19.0.0
